Extract the authorization-server routes into router.ts
Move the seven hosted-OAuth-proxy endpoints (RFC 8414 AS metadata x2, RFC 7591
register, and the authorize/consent/callback/token flow) plus their helpers
(errorReason, readAuthorizeQuery, serializeAuthorizeQuery, redirectHostOf,
setTxnCookie, resolveClient) out of the transport's buildApp into a new
mountAuthorizationServer() under src/auth/authorizationServer/, where the rest of
the AS code lives (per AGENTS.md). buildApp now calls it with a small deps object.

Verbatim move — routes and their per-endpoint 404 guards are unchanged, so
behaviour is preserved; the end-to-end proxy suite drives every moved route.
streamableHttp.ts drops from 1331 to ~1000 lines.
